refactor(geom): Split tests, add Greiner-Hormann, fix primitives
- Split `geom_algorithms_test.cc` into multiple translation units to reduce compilation memory usage and updated `Tests/CMakeLists.txt`.
- Implement Greiner-Hormann algorithm in `BooleanPolygonOperations` to correctly handle concave polygons for intersection, union, and difference, replacing the previous approximation.
- Fix `Triangle::contains` to correctly handle clockwise winding.
- Update `Segment::slope` to return infinity for vertical segments.
- Add validation for positive radii in `Ellipse` and `RotatedEllipse`.
- Handle collinear triples in `ConvexPolygonOffset` and add safety loop counter in `RotatingCalipers`.
- Use `std::mt19937` for randomization in `DelaunayTriangulationRandomizedIncremental`.
- Add degeneracy checks in `Triangle3D` barycentric coordinates and `BruteForceConvexHull`.
